Add module for creating a read policy
We generate policy JSON to read each secret individually, which requires
creating a policy and merging the combined JSON into it. This creates
some additional work and also sometimes creates IAM policy statements
which are too large for a single policy.

This creates a module which will create a standalone policy for reading
a list of secrets, generating a much smaller policy which would not
exceed the quota until an extreme number of secrets is needed. It can
also attach it to roles by name.