Add module for user-managed secrets
On most applications we have at least one secret that must be set
manually by a developer. This adds a module to directly support this use
case.

It better supports these secrets by disabling the rotation policy and
role (since rotation will not be enabled) and making it easier to
provide the default blank value using environment variable format.